⚡️Target ES2020 in produced bundle
The bundle of fast-check that we have been producing until now were targeting ES2017. As such any feature not compatible with this standard got transpiled to be compatible.

While allowing older versions of node to be able to run fast-check properly it created a performance cliff for others. With v4 and our target being moved to ES2020 we increased the target version to ES2020.