🔧(vitest) Update tsconfig to common config
The main impact is that object spreading will not be polyfilled anymore at build time. In other words any runtime of JavaScrip not supporting the syntax `{...obj}` will break.

Such feature has been supported for years so we don't expect any major blocker to move to the release including this change. That said it's still a breaking change.