👷 Get rid of LFS storage
Our LFS limits are constantly breaching. Part of the problem is possibly that the repository is not only pulled by my CI or myself but by many others. As such LFS limits can easily get breached.

This PR consists into dropping everything from LFS and use another repository for assets so that we don't clutter this one with huge amounts of binary data.