✨ Introduce a revamped `waitAll` for `scheduler`
In version 4.0.0, we decided to make `waitAll` way more touchy by restricting the amount of extra works it can perform. Our initial version was performing too many extra works but not enough in some cases. As such it made it hardly predictable.

With our first few minors of version 4.x, we started to pave the way towards more predictable awaiters. Our new rebranded `waitAll`, no called `waitIdle`, relies on these new primitives to provide a more predictable and reliable waiting mechanism. The naming has been changed to avoid confusion as it does not wait "all" tasks to finish, but rather waits for the ones it can detect.

Our next steps will consists into deciding about the future of `waitOne` and `waitAll`. As such none of them makes sense for a scheduler and we should probably not make them as simple to access. Our current plan would be to drop `waitOne`, `waitAll` and `count` but offer users with a way to rebuild `waitAll` by calling `waitIdle` with an extra option like `{immediate: true}`.