✨ Introduce `waitIdle`, a revamped `waitAll` for `scheduler` (#6016)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

In version 4.0.0, we decided to make `waitAll` way more touchy by
restricting the amount of extra works it can perform. Our initial
version was performing too many extra works but not enough in some
cases. As such it made it hardly predictable.

With our first few minors of version 4.x, we started to pave the way
towards more predictable awaiters. Our new rebranded `waitAll`, no
called `waitIdle`, relies on these new primitives to provide a more
predictable and reliable waiting mechanism. The naming has been changed
to avoid confusion as it does not wait "all" tasks to finish, but rather
waits for the ones it can detect.

Our next steps will consists into deciding about the future of `waitOne`
and `waitAll`. As such none of them makes sense for a scheduler and we
should probably not make them as simple to access. Our current plan
would be to drop `waitOne`, `waitAll` and `count` but offer users with a
way to rebuild `waitAll` by calling `waitIdle` with an extra option like
`{immediate: true}`.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ✨ Introduce new features
- [x] Impacts: None, just adding features

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->