🗑️ Deprecate `waitOne` and `waitAll` (#6026)
**Description**

The methods `waitOne` and `waitAll` suffer from being hardly usable
without a deep knowledge of their way to work. In a code snippet like:

```js
const f = s.scheduleFunction(something);
async function abc() {
  const a = await somethingAlreadyResolvedFromCache;
  const b = await somethingElseAlreadyResolvedFromCache;
  await f(a, b);
}
const p = abc();
await s.waitAll();
```

The `waitAll` call will be unable to wait for `p` to be done. It will
end even before `f` got called as it has been developed to avoid letting
any promise be executed before itself. Before 4.2.0, the only way to
wait safely for `p` to be done with everything leading to `p` being
scheduled was to use `waitFor(p)`.

Starting at 4.2.0, you can now use `waitIdle`. It will really wait for
everything needing to be scheduled to be so.

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `pnpm run bump` and flag
the impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 🗑️ Deprecate
- [x] Impacts: None yet

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->