💚 Skip expensive CI checks on Windows runner
Windows runners used to be unreliable for the tests on Docs. As such while being useful to make sure we generate our documentation, the drawback of having them making the CI unstable due to Windows runners was not an acceptable status.

This PR stops running these tests against Windows as they proved unreliable lately and led to lots of frustration on the CI.