🐛(vitest) No shallow diff in Vitest, full diff by default
When running our Vitest plugin against v3 of fast-check some users might suffer from the fact that the diff is shallow by default. This is not be the case of the dafult setup when relying on our plugin.

As such this PR makes sure that the default will avoid it.

Fixes #5995